Move around validators
